Fix size_t underflow in binary search in x509cert processing
The original code used `size_t` for `first`,`last` & `middle` sentinels
in a binary search algorithm when looking up OIDs in a table in the
OidToNameLookup & OidToKeyUsageValueLookup functions.

If an OID which came before all of the entries in the table
was looked up eventually the `last` sentinel would be decreased
to -1 and the search should have ended, as the condition `first <= last`
should have failed. However, since an unsigned type (`size_t`) was used
for the sentinels, decreasing `last` to -1 underflowed it and the
code subsequently crashed.