Rename the misleadingly named method GetSubjecPublicKeyAlgorithm.
The method name indicated that it would return the algorithm of the
public key that is signed by the certificate. Instead it returned
the algorithm used to sign the public key. The current commit changes
it return the OID of the public key algorithm and adds a new method,
GetSubjectSignatureAlgorithm, to retrieve the information that was
originally returned --- the signature algorithm.